requirements.txt
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
notion-client>=2.0
python-dateutil>=2.8
anthropic>=0.30
